Refactor artefact registry and async loading
Major refactor of artefact/symbol handling and plugin loading.

- Artefact factory rename: replaced artefactProtoFactory with artefactFactory in GAML and updated callers to use GAML.getArtefactFactory()/registerArtefactProtoFactory accordingly.
- ArtefactRegistry overhaul: simplified and renamed internal maps (statement/variable artefacts), removed old Guava multimaps, introduced lazy cached facets, clarified constants and javadoc, changed APIs to consistently treat statement vs variable artefacts and adjusted related helpers.
- AbstractGamlAdditions: switched to use artefact factory create methods, simplified keyword handling for symbols and facets, removed some legacy registry calls.
- GamaBundleLoader: plugin addition loading moved to asynchronous execution (CompletableFuture.runAsync) using virtual threads for the executor; timing log messages adjusted.
- ISymbolKind: use EnumSet for classification (variables/statements), renamed methods/fields for clarity.
- IExpressionDescriptionFactory: added accessors for true/false/null expressions.
- ISymbol: changed getOrder return type to long and adjusted docs.

Purpose: modernize and simplify artefact registration and lookup, improve thread-safety and startup responsiveness by loading plugin extensions asynchronously, and clean up APIs and naming for clarity.

+++ gama.api/src/gama/api/gaml/symbols/ISymbol.java
@@ -27,8 +27,8 @@ import gama.api.utils.interfaces.INamed;
  * </p>
  * 
  * <p>
- * Symbols form a hierarchical structure where symbols can contain child symbols (e.g., a species contains variables
- * and actions). They are ordered to maintain declaration order and support proper initialization sequences.
+ * Symbols form a hierarchical structure where symbols can contain child symbols (e.g., a species contains variables and
+ * actions). They are ordered to maintain declaration order and support proper initialization sequences.
  * </p>
  * 
  * <h2>Key Responsibilities</h2>
@@ -57,7 +57,7 @@ public interface ISymbol extends INamed, IDisposable {
 	 *
 	 * @return the declaration order (lower values indicate earlier declaration)
 	 */
-	int getOrder();
+	long getOrder();
 
 	/**
 	 * Sets the declaration order of this symbol.
