+++ gama.api/src/gama/api/kernel/skill/IArchitecture.java
@@ -18,13 +18,13 @@ import gama.api.runtime.scope.IScope;
 
 /**
  * The Interface IArchitecture.
- * 
+ *
  * <p>
  * Defines the control architecture that governs how an agent executes its behaviors during a simulation step. An
  * architecture determines the order and conditions under which an agent's reflexes and other behaviors are executed,
  * allowing for different behavioral paradigms (reactive, cognitive, BDI, FSM, etc.).
  * </p>
- * 
+ *
  * <h3>Purpose</h3>
  * <p>
  * Architectures provide:
@@ -36,7 +36,7 @@ import gama.api.runtime.scope.IScope;
  * <li><b>Validation:</b> Verify behaviors are compatible with the architecture</li>
  * <li><b>Pre-processing:</b> Setup before population-level execution</li>
  * </ul>
- * 
+ *
  * <h3>Built-in Architectures</h3>
  * <ul>
  * <li><b>reflex (default):</b> Execute all reflexes in declaration order</li>
@@ -47,17 +47,17 @@ import gama.api.runtime.scope.IScope;
  * <li><b>user_first:</b> User command first, then reflexes</li>
  * <li><b>user_last:</b> Reflexes first, then user command</li>
  * </ul>
- * 
+ *
  * <h3>Extension Architectures</h3>
  * <ul>
  * <li><b>bdi:</b> Belief-Desire-Intention for cognitive agents (extension)</li>
  * <li><b>moving:</b> Path-following and navigation behaviors</li>
  * </ul>
- * 
+ *
  * <h3>Usage in GAML</h3>
- * 
+ *
  * <h4>1. Default Reflex Architecture</h4>
- * 
+ *
  * <pre>
  * <code>
  * species animal {
@@ -68,21 +68,21 @@ import gama.api.runtime.scope.IScope;
  * }
  * </code>
  * </pre>
- * 
+ *
  * <h4>2. Finite State Machine</h4>
- * 
+ *
  * <pre>
  * <code>
  * species robot control: fsm {
  *     state searching initial: true {
  *         transition to: moving when: target != nil;
  *     }
- *     
+ *
  *     state moving {
  *         do goto target: target;
  *         transition to: working when: location = target;
  *     }
- *     
+ *
  *     state working {
  *         do work;
  *         transition to: searching when: task_complete;
@@ -90,103 +90,103 @@ import gama.api.runtime.scope.IScope;
  * }
  * </code>
  * </pre>
- * 
+ *
  * <h4>3. Weighted Tasks</h4>
- * 
+ *
  * <pre>
  * <code>
  * species worker control: weighted_tasks {
  *     task eat priority: hunger_level {
  *         // Higher hunger = higher priority
  *     }
- *     
+ *
  *     task work priority: 5 {
  *         // Fixed priority
  *     }
- *     
+ *
  *     task rest priority: 100 - energy {
  *         // Lower energy = higher priority
  *     }
  * }
  * </code>
  * </pre>
- * 
+ *
  * <h4>4. BDI Architecture (Extension)</h4>
- * 
+ *
  * <pre>
  * <code>
  * species cognitive_agent control: bdi {
  *     perception {
  *         // Update beliefs based on environment
  *     }
- *     
+ *
  *     rule when: has_belief("food_nearby") {
  *         desire: "eat_food";
  *     }
- *     
+ *
  *     plan eat_food intention: "eat_food" {
  *         // Execute plan to achieve intention
  *     }
  * }
  * </code>
  * </pre>
- * 
+ *
  * <h4>5. User-Controlled</h4>
- * 
+ *
  * <pre>
  * <code>
  * species player control: user_only {
  *     // No automatic behaviors
  *     // Agent only acts when user issues commands
- *     
+ *
  *     user_command "Move North" {
  *         location <- location + {0, 1};
  *     }
- *     
+ *
  *     user_command "Attack" {
  *         do attack;
  *     }
  * }
  * </code>
  * </pre>
- * 
+ *
  * <h3>Java Usage - Implementing Custom Architecture</h3>
- * 
+ *
  * <pre>
  * <code>
  * {@literal @}skill(name = "custom_architecture")
  * public class CustomArchitecture extends AbstractArchitecture {
- *     
+ *
  *     {@literal @}Override
  *     public Object privateExecuteIn(IScope scope) throws GamaRuntimeException {
  *         IAgent agent = scope.getAgent();
- *         
+ *
  *         // Custom execution logic
  *         // 1. Pre-processing
  *         // 2. Execute behaviors based on custom rules
  *         // 3. Post-processing
- *         
+ *
  *         return null;
  *     }
- *     
+ *
  *     {@literal @}Override
  *     public boolean init(IScope scope) {
  *         // Initialize architecture for agent
  *         return true;
  *     }
- *     
+ *
  *     {@literal @}Override
  *     public boolean abort(IScope scope) {
  *         // Clean up when agent dies
  *         return true;
  *     }
- *     
+ *
  *     {@literal @}Override
  *     public void verifyBehaviors(ISpecies context) {
  *         // Validate that behaviors are compatible
  *         // Throw GamaRuntimeException if invalid
  *     }
- *     
+ *
  *     {@literal @}Override
  *     public void preStep(IScope scope, IPopulation&lt;?&gt; population) {
  *         // Population-level pre-processing before stepping agents
@@ -194,7 +194,7 @@ import gama.api.runtime.scope.IScope;
  * }
  * </code>
  * </pre>
- * 
+ *
  * <h3>Architecture Lifecycle</h3>
  * <ol>
  * <li><b>Compilation:</b> verifyBehaviors() validates behaviors against architecture</li>
@@ -204,14 +204,14 @@ import gama.api.runtime.scope.IScope;
  * <li><b>Execution:</b> privateExecuteIn() called for each agent each cycle</li>
  * <li><b>Termination:</b> abort() called when agent dies</li>
  * </ol>
- * 
+ *
  * <h3>Design Patterns</h3>
  * <ul>
  * <li><b>Strategy Pattern:</b> Different architectures = different behavior execution strategies</li>
  * <li><b>Template Method:</b> AbstractArchitecture provides template, subclasses customize</li>
  * <li><b>Command Pattern:</b> Behaviors encapsulated as executable commands</li>
  * </ul>
- * 
+ *
  * <h3>Implementation Notes</h3>
  * <ul>
  * <li>Implements both ISkill (capabilities) and IStatement (executable)</li>
@@ -220,7 +220,7 @@ import gama.api.runtime.scope.IScope;
  * <li>Order can be customized (default returns 0)</li>
  * <li>preStep allows population-wide setup before individual agent execution</li>
  * </ul>
- * 
+ *
  * @see ISkill
  * @see AbstractArchitecture
  * @see IAgent
@@ -276,7 +276,7 @@ public interface IArchitecture extends ISkill, IStatement {
 	 * @return the order
 	 */
 	@Override
-	default int getOrder() { return 0; }
+	default long getOrder() { return 0; }
 
 	/**
 	 * Sets the order.
