+++ gama.api/src/gama/api/utils/tests/AbstractSummary.java
@@ -38,7 +38,7 @@ public abstract class AbstractSummary<S extends WithTestSummary<?>> implements I
 	private final String title;
 
 	/** The index. */
-	public final int index = COUNTER.GET_UNIQUE();
+	public final long index = COUNTER.GET_UNIQUE();
 
 	/** The error. */
 	public String error;
@@ -79,7 +79,7 @@ public abstract class AbstractSummary<S extends WithTestSummary<?>> implements I
 	 *
 	 * @return the index
 	 */
-	public int getIndex() { return index; }
+	public long getIndex() { return index; }
 
 	/**
 	 * Gets the uri.
